Unify RectanglesScreen.cs across MonoGame/raylib/SilkNet samples
Converges the three drifted copies into one shared file (MonoGameGumInCode
is canonical, linked into raylib/SilkNet via <Compile Include Link>),
following the CirclesScreen (#4024) pattern.

- Adds Antialiasing and Blend rows to MG (already fully supported by
  RectangleRuntime, just missing from the sample).
- Adds Per-corner radii row to raylib (CustomRadius* round-trips there but
  isn't rendered yet -- left ungated so the gap is visible, same policy as
  Circles' Blend/Antialiasing gaps).
- Converts raylib's split Rotation(filled)/Rotation(outline-unsupported)
  rows into the single BuildRotationRow(bool) used by MG/SilkNet, matching
  how CirclesScreen handles raylib's lack of gradient-on-stroke (render a
  plain outline instead of swapping in an "unsupported" label).
- Drops a stale raylib-only dropshadow diagnostic cell (alpha=32, "This
  seems wrong" comment) that predates the #2865 shadow-blur fix and should
  have been removed alongside 43a755b5e.
- Standardizes on new Color(r, g, b, a) everywhere (no named-color statics)
  and shortens section labels to match CirclesScreen's terser style.

Closes #4029
